requirements.txt
alembic==1.4.2
attrs==19.3.0
beautifulsoup4==4.9.1
click==7.1.2
Flask==1.1.2
Flask-Cors==3.0.8
flask-marshmallow==0.13.0
Flask-Migrate==2.5.2
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.1
gunicorn==20.0.4
importlib-metadata==1.7.0
iniconfig==1.0.1
itsdangerous==1.1.0
Jinja2==2.11.2
Mako==1.1.3
MarkupSafe==1.1.1
marshmallow==3.7.1
marshmallow-sqlalchemy==0.23.1
more-itertools==8.4.0
packaging==20.4
pluggy==0.13.1
py==1.9.0
pyparsing==2.4.7
pytest==6.0.1
python-dateutil==2.8.1
python-editor==1.0.4
PyYAML==5.3.1
six==1.15.0
soupsieve==2.0.1
SQLAlchemy==1.3.18
swagger-ui-py==0.3.0
toml==0.10.1
waitress==1.4.4
WebOb==1.8.6
WebTest==2.0.35
Werkzeug==1.0.1
zipp==3.1.0
psycopg2==2.8.5
